+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [master]
+> 2017-09-06
+
+- **Chore:** TextField support number type value
+
+[master]: https://github.com/yyssc/ssc-grid/compare/v0.69.14...master
+
 ## [v0.69.14]
 > 2017-09-06