+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 ## [master]
 > 2017-09-06
 
+[master]: https://github.com/yyssc/ssc-grid/compare/v0.69.15...master
+
+## [v0.69.15]
+> 2017-09-06
+
 - **Chore:** TextField support number type value
 - **Feature:** ValidateFloatInput support validation message
 - **Feature:** 支持从组件外通过ref调用`doValid()`方法
 - **Bug:** 解决不能传值的bug
 
-[master]: https://github.com/yyssc/ssc-grid/compare/v0.69.14...master
+[v0.69.15]: https://github.com/yyssc/ssc-grid/compare/v0.69.14...v0.69.15
 
 ## [v0.69.14]
 > 2017-09-06